+++ R/assess_db.R
@@ -1,8 +1,7 @@
-
-
 #' Read in the raw assessment database from the NYSED website
 #'
-#' @param end_year
+#' @param end_year 4 digit integer representing the end of the desired school
+#' year.  eg, 2014-15 is 2015.
 #'
 #' @return data frame with assessment results, by school
 #' @export
@@ -15,21 +14,10 @@ get_raw_assess_db <- function(end_year) {
     'yr2013' = 'http://data.nysed.gov/files/assessment/3-8-2013-14.zip'
   )
   assess_url <- assess_urls[[paste0('yr', end_year)]]
+  local_files <- zip_to_temp(assess_url)
+  assess <- extract_mdb(local_files)
 
-  #download and unzip
-  tname <- tempfile(pattern = "enr", tmpdir = tempdir(), fileext = ".zip")
-  tdir <- tempdir()
-  downloader::download(assess_url, dest = tname, mode = "wb")
-  utils::unzip(tname, exdir = tdir)
-
-  #read file
-  assess_files <- utils::unzip(tname, exdir = ".", list = TRUE)
-  mdb_file <- assess_files[grepl('.mdb', assess_files$Name, fixed = TRUE), ]$Name
-  mdb_file <- file.path(tdir, mdb_file)
-  file.rename(mdb_file, file.path(tdir, 'assess_data.mdb'))
-
-  assess <- Hmisc::mdb.get(file = file.path(tdir, 'assess_data.mdb'))
-
+  #2014 data file included both the 2013 data and the 2014 data. process.
   if (end_year == 2013) {
     out <- assess$`3-8_ELA_AND_MATH_REPORT_FOR_RELEASE_2013`
   } else if (end_year == 2014) {