+++ NEWS.md
@@ -1,9 +1,12 @@
-# nyschooldata 0.0.2
+# nyschooldata 0.0.3
+* refactored unzip/download code into separate functions in `util.R`, with tests
+* removed `downloader` dependency
+* added `rc_db.R` functions (`get_raw_rc_db()`), with tests
 
+# nyschooldata 0.0.2
 * added wercker CI files
 * documentation and build status
 
 # nyschooldata 0.0.1
-
 * initial commits
 * assessment database, 2013-2015, using helper in Hmisc